+++ cmd/influxd/config.go
@@ -72,12 +72,11 @@ type Config struct {
 	Graphites []Graphite `toml:"graphite"`
 	Collectd  Collectd   `toml:"collectd"`
 
-	InputPlugins struct {
-		UDPInput struct {
-			Enabled bool `toml:"enabled"`
-			Port    int  `toml:"port"`
-		} `toml:"udp"`
-	} `toml:"input_plugins"`
+	UDP struct {
+		Enabled     bool   `toml:"enabled"`
+		BindAddress string `toml:"bind-address"`
+		Port        int    `toml:"port"`
+	} `toml:"udp"`
 
 	Broker struct {
 		Port    int      `toml:"port"`
@@ -169,9 +168,14 @@ func NewConfig() *Config {
 	return c
 }
 
-// DataAddr returns the binding address the data server
+// DataAddr returns the TCP binding address for the data server.
 func (c *Config) DataAddr() string {
 	return net.JoinHostPort(c.BindAddress, strconv.Itoa(c.Data.Port))
+}
+
+// DataAddrUDP returns the UDP address for the series listener.
+func (c *Config) DataAddrUDP() string {
+	return net.JoinHostPort(c.UDP.BindAddress, strconv.Itoa(c.UDP.Port))
 }
 
 // DataURL returns the URL required to contact the data server.