+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bugfixes
 - [#1836] (https://github.com/influxdb/influxdb/pull/1836): Store each parsed shell command in history file.
+- [#1789] (https://github.com/influxdb/influxdb/pull/1789): add --config-files option to fpm command. Thanks @kylezh
  
 ### Features